Make 4.5 branch buildable in Azure Pipelines (#2714)

+++ scripts/cibuild/ConfigureVstsBuild.ps1
@@ -29,12 +29,8 @@ param
 
     [Parameter(Mandatory=$True)]
     [string]$BuildRTM,
-
-    [Parameter(Mandatory=$True)]
-    [string]$FunctionalTestBuildId,
-
-    [Parameter(Mandatory=$True)]
-    [string]$VstsRestApiRootUrl
+    
+    [switch]$SkipUpdateBuildNumber
 )
 
 Function Get-Version {
@@ -70,6 +66,7 @@ Function Update-VsixVersion {
     # Evaluate the new version
     $newVersion = Get-Version $ReleaseProductVersion $buildNumber
     Write-Host "Updating the VSIX version [$oldVersion] => [$newVersion]"
+    Write-Host "##vso[task.setvariable variable=VsixBuildNumber;]$newVersion"
     # setting the revision to the new version
     $root.Metadata.Identity.Version = "$newVersion"
 
@@ -89,6 +86,8 @@ $regKeyNuGet = "HKLM:SOFTWARE\Microsoft\StrongName\Verification\*,31bf3856ad364e
 $regKeyNuGet32 = "HKLM:SOFTWARE\Wow6432Node\Microsoft\StrongName\Verification\*,31bf3856ad364e35"
 
 $has32bitNode = Test-Path "HKLM:SOFTWARE\Wow6432Node"
+$regKeyFileSystem = "HKLM:SYSTEM\CurrentControlSet\Control\FileSystem"
+$enableLongPathSupport = "LongPathsEnabled"
 
 # update submodule NuGet.Build.Localization
 $NuGetClientRoot = $env:BUILD_REPOSITORY_LOCALPATH
@@ -125,47 +124,66 @@ if (-not (Test-Path $regKeyNuGet) -or ($has32bitNode -and -not (Test-Path $regKe
     }
 }
 
+if (-not (Test-Path $regKeyFileSystem)) 
+{
+	Write-Host "Enabling long path support on the build machine"
+	Set-ItemProperty -Path $regKeyFileSystem -Name $enableLongPathSupport -Value 1
+}
+
 
 if ($BuildRTM -eq 'true')
 {
     # Set the $(NupkgOutputDir) build variable in VSTS build
     Write-Host "##vso[task.setvariable variable=NupkgOutputDir;]ReleaseNupkgs"
     Write-Host "##vso[task.setvariable variable=VsixPublishDir;]VS15-RTM"
-    $numberOfTries = 0
-    do{
-        Write-Host "Waiting for buildinfo.json to be generated..."
-        $numberOfTries++
-        Start-Sleep -s 15
-    }
-    until ((Test-Path $BuildInfoJsonFile) -or ($numberOfTries -gt 50))
-    $json = (Get-Content $BuildInfoJsonFile -Raw) | ConvertFrom-Json
-    $currentBuild = [System.Decimal]::Parse($json.BuildNumber)
-    # Set the $(Revision) build variable in VSTS build
-    Write-Host "##vso[task.setvariable variable=Revision;]$currentBuild"
-    Write-Host "##vso[build.updatebuildnumber]$currentBuild" 
-    $oldBuildOutputDirectory = Split-Path -Path $BuildInfoJsonFile
-    $branchDirectory = Split-Path -Path $oldBuildOutputDirectory
-    $newBuildOutputFolder =  Join-Path $branchDirectory $currentBuild
-    if(Test-Path $newBuildOutputFolder)
+    # Only for backward compatibility with orchestrated builds
+    if(-not $SkipUpdateBuildNumber)
     {
-        Move-Item -Path $BuildInfoJsonFile -Destination $newBuildOutputFolder
-        Remove-Item -Path $oldBuildOutputDirectory -Force
-    }
-    else
-    {
-        Rename-Item $oldBuildOutputDirectory $currentBuild
-    }
-    
+        $numberOfTries = 0
+        do{
+            Write-Host "Waiting for buildinfo.json to be generated..."
+            $numberOfTries++
+            Start-Sleep -s 15
+        }
+        until ((Test-Path $BuildInfoJsonFile) -or ($numberOfTries -gt 50))
+        $json = (Get-Content $BuildInfoJsonFile -Raw) | ConvertFrom-Json
+        $currentBuild = [System.Decimal]::Parse($json.BuildNumber)
+        # Set the $(Revision) build variable in VSTS build
+        Write-Host "##vso[task.setvariable variable=Revision;]$currentBuild"
+        Write-Host "##vso[build.updatebuildnumber]$currentBuild" 
+        $oldBuildOutputDirectory = Split-Path -Path $BuildInfoJsonFile
+        $branchDirectory = Split-Path -Path $oldBuildOutputDirectory
+        $newBuildOutputFolder =  Join-Path $branchDirectory $currentBuild
+        if(Test-Path $newBuildOutputFolder)
+        {
+            Move-Item -Path $BuildInfoJsonFile -Destination $newBuildOutputFolder
+            Remove-Item -Path $oldBuildOutputDirectory -Force
+        }
+        else
+        {
+            Rename-Item $oldBuildOutputDirectory $currentBuild
+        }   
+    }    
 }
 else
 {
-    $revision = Get-Content $BuildCounterFile
-    $newBuildCounter = [System.Decimal]::Parse($revision)
-    $newBuildCounter++
-    Set-Content $BuildCounterFile $newBuildCounter
-    # Set the $(Revision) build variable in VSTS build
-    Write-Host "##vso[task.setvariable variable=Revision;]$newBuildCounter"
-    Write-Host "##vso[build.updatebuildnumber]$newBuildCounter"
+    # Only for backward compatibility with orchestrated builds
+    if(-not $SkipUpdateBuildNumber)
+    {
+        $revision = Get-Content $BuildCounterFile
+        $newBuildCounter = [System.Decimal]::Parse($revision)
+        $newBuildCounter++
+        Set-Content $BuildCounterFile $newBuildCounter
+        # Set the $(Revision) build variable in VSTS build
+        Write-Host "##vso[task.setvariable variable=Revision;]$newBuildCounter"
+        Write-Host "##vso[build.updatebuildnumber]$newBuildCounter"
+        Write-Host "##vso[task.setvariable variable=BuildNumber;isOutput=true]$newBuildCounter"
+    }
+    else
+    {
+        $newBuildCounter = $env:BUILD_BUILDNUMBER        
+    }
+
     $jsonRepresentation = @{
         BuildNumber = $newBuildCounter
         CommitHash = $env:BUILD_SOURCEVERSION
@@ -182,5 +200,5 @@ else
         Write-Error "Failed to get product version."
         exit 1
     }
-    Update-VsixVersion -manifestName source.extension.vs15.vsixmanifest -ReleaseProductVersion $productVersion -buildNumber $newBuildCounter
+    Update-VsixVersion -manifestName source.extension.vs15.vsixmanifest -ReleaseProductVersion $productVersion -buildNumber $env:BUILDNUMBER
 }
